sync-dotfiles: Release 0.1.6-rc
[Optimizations]
Reduced more string allocations

[Refactoring]
Documented the code and added comments to make it easier to understand.

[Fixes]
Safely exit the process after performing a task.
Fix a bug where if the user specifies a config path with a trailing slash, the program will crash.
Additionally checks if the config with the same name already exists while inserting a new config.

[Features]

Change-Id: I59a3bd774ee1285306d364540e1a40849c54d73d
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>