sync-dotfiles: Release 0.1.6
[Optimizations]
Reduced more string allocations

[Refactoring]
Documented the code and added comments to make it easier to understand.
Added parsing of $(pwd) prefix in the config path.
Added a stripped down version of `hristogochev/merkle_hash` for hashing the config files.
Added another field to Config structure to hold the config type. `Dir or File`.

[Fixes]
Safely exit the process after performing a task.
Fix a bug where if the user specifies a config path with a trailing slash, the program will crash.
Additionally checks if the config with the same name already exists while inserting a new config.
Fixed push and pull of a single config file. `an example can be your .zshrc file`.

[Features]

Change-Id: I9a26757ef99766162f058ed8c8d9a011e7c7cc56
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>